remove CONAN_V2_MODE env-var and compiler_id detection
